Move useSnackbar into components/Snackbar folder. Remove custom offset story. Revert story name to Snackbar
